# BackDoorLogic: optimize persistDataBundle method #2258