… src/main/java/teammates/logic/backdoor/BackDoorOperation.java
… src/main/java/teammates/logic/backdoor/BackDoorServlet.java
… src/main/java/teammates/storage/api/AccountsDb.java
… src/main/java/teammates/storage/api/EntitiesDb.java
… src/main/java/teammates/storage/api/FeedbackQuestionsDb.java
… src/main/java/teammates/storage/api/InstructorsDb.java
… src/main/java/teammates/ui/controller/AdminInstructorAccountAddAction.java
… src/test/java/teammates/test/cases/search/InstructorSearchTest.java
… src/test/java/teammates/test/driver/BackDoor.java